+++ CHANGELOG.md
@@ -8,6 +8,7 @@ CHANGELOG ekc-tournament
 # Version 1.2.14
 [x] Additional attribute on teams: sports club / city (optional column for registration table)
 [x] Removed attributes 'camping count' and 'breakfast count' from teams overview table
+[x] Swiss System: save single result: mark missing results
 
 # Version 1.2.13
 [x] Fix timer: off by one minute